[web\schema] merge schema to module
